+++ hil-test/tests/aes_dma.rs
@@ -6,11 +6,7 @@
 #![no_main]
 
 use esp_hal::{
-    aes::{
-        dma::{CipherMode, WithDmaAes},
-        Aes,
-        Mode,
-    },
+    aes::{dma::CipherMode, Aes, Mode},
     dma::{Dma, DmaPriority},
     dma_buffers,
     peripherals::Peripherals,